feat: api, hook
