+++ src/components/Filter.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import '@/components/Filter.css';
 import { useQuery } from '@tanstack/react-query';
-import { fetchPerformances, IPerformancePayload } from '@/apis/Performances';
+import { usePerformances } from '@/hooks/usePerformances';
 
 //임의의 카테고리 더미 데이터, api 연결되면 지울 예정
 export const category: string[] = [
@@ -25,21 +25,9 @@ const Filter: React.FC = () => {
   const [selectedCategory, setSelectedCategory] = useState<string>('전체');
   const [searchTerm, setSearchTerm] = useState('');
 
-  const {
-    /** data : 쿼리 함수의 반환값을 저장하는 변수 */
-    data: performances,
-    /** isLoading : 쿼리 함수가 아직 반환되지 않았을 때 true를 반환하는 변수 */
-    isLoading,
-    /** isError : 쿼리 함수가 에러를 발생시킬 경우 true를 반환하는 변수 */
-    isError,
-    /** refetch : 쿼리 함수를 다시 실행하는 함수 (재조회 용으로 이용됨) */
-    refetch,
-  } = useQuery({
-    /** queryKey : 향후 캐싱 기능에서 이용이 되며, 중복되는 요청을 방지할때에 구분자로도 이용됨 */
-    queryKey: ['performances'],
-    /** queryFn : 실행할 함수를 선언해주며, 주로 API요청을 호출하는 로직으로 이용됨. */
-    queryFn: fetchPerformances,
-  });
+  // const {performance} = usePerformances();
+
+  
 
   const handleClickContainer = () => {
     setDropdownView(!isDropdownView);
