+++ src/firebase.tsx
@@ -4,7 +4,6 @@ import 'firebase/auth';
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
-export const BACKEND_URL = import.meta.env.VITE_BACKEND_URL;
 
 // Your web app's Firebase configuration
 const firebaseConfig = {
